23th
终于是把输入跟计时器做完了 明天做render系统和resource管理器

+++ include/core/Timer.h
@@ -17,7 +17,7 @@ public:
 
 	float Get_framecount()const;
 
-	float Get_fps()const;
+	unsigned int Get_fps()const;
 
 	void Reset();
 private:
